Update setMaskFromFootprintList to match afw trunk; changes to make algorithms compile cleanly with gcc 4.1.2

+++ include/lsst/meas/algorithms/CR.h
@@ -14,7 +14,7 @@ template <typename MaskedImageT>
 std::vector<lsst::afw::detection::Footprint::Ptr>
 findCosmicRays(MaskedImageT& image,
                PSF const &psf,
-               float const bkgd,
+               double const bkgd,
                lsst::pex::policy::Policy const& policy,
                bool const keep = false
               );
